refactor: :truck: move interfaces to src
