+++ src/components/Layout/Layout.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 import { Container } from '@mantine/core'
 import CustomHeader from '../CustomHeader/CustomHeader'
 import Footer from '../CustomFooter/CustomFooter'
-import { ILayoutProps } from '../../ts/interfaces/Layout.interface'
+import { ILayoutProps } from '../../interfaces/Layout.interface'
 import { navLinks } from '../CustomHeader/navLinks'
 
 const Layout: FC<ILayoutProps> = ({ children }) => (
