+++ src/interfaces/Article.interface.ts
