+++ src/interfaces/Bookmark.interface.ts
