+++ src/interfaces/CustomImage.interface.ts
