+++ src/interfaces/NowPlaying.type.ts
